feat: add Beta Factor framework, remove max-turns limit, cleanup dashboard
- Beta Factor system: BetaSnapshot, BetaReview, BetaInsight models with
  4-stage feedback loop (capture → review → aggregate → scorecard)
- Beta scorecard API for AI-assisted non-technical factor analysis
- Auto-aggregate beta insights in daily scheduler
- Manual "Beta聚合" button in UI
- Remove --max-turns from all Claude CLI calls (use budget limit only)
- Delete legacy Streamlit dashboard (src/dashboard/)
- Add mass rebacktest and migration scripts
- Various bot trading fixes (exit_config jsonb, execution price logic)

+++ api/services/regime_service.py
@@ -129,11 +129,11 @@ def compute_weekly_regimes(
             "week_start": current_monday,
             "week_end": current_friday,
             "regime": regime.regime,
-            "confidence": round(regime.confidence, 4),
-            "trend_strength": round(regime.trend_strength, 4),
-            "volatility": round(regime.volatility, 4),
-            "breadth": round(regime.breadth, 4),
-            "index_return_pct": round(index_return, 4),
+            "confidence": float(round(regime.confidence, 4)),
+            "trend_strength": float(round(regime.trend_strength, 4)),
+            "volatility": float(round(regime.volatility, 4)),
+            "breadth": float(round(regime.breadth, 4)),
+            "index_return_pct": float(round(index_return, 4)),
         })
 
         current_monday += timedelta(days=7)
